requirements.txt
asgiref==3.8.1
Django==5.1.2
psycopg2==2.9.10
python-decouple==3.8
sqlparse==0.5.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
